+++ prost-reflect/src/dynamic/serde/de/kind.rs
@@ -547,6 +547,17 @@ impl<'a, 'de> Visitor<'de> for MessageVisitorInner<'a> {
                 if let Some(value) =
                     map.next_value_seed(OptionalFieldDescriptorSeed(&field, self.1))?
                 {
+                    if let Some(oneof_desc) = field.containing_oneof() {
+                        for field in oneof_desc.fields() {
+                            if self.0.has_field(field.number()) {
+                                return Err(Error::custom(format!(
+                                    "multiple fields provided for oneof '{}'",
+                                    oneof_desc.name()
+                                )));
+                            }
+                        }
+                    }
+
                     self.0.set_field(field.number(), value);
                 }
             } else if self.1.deny_unknown_fields {